stitchlabos/image/src/modules/live-jogd/filesystem/home/pi/live_jogd/requirements.txt
pyserial>=3.5
aiohttp>=3.8
